Add request/response processor concepts to axion
Idea here is to make it possible to process both
request and response in type safe manner. Hence `TypeVar`.
With it, a plugin receives a `RQ` (request) and process it.
It has to return an axion's version of a `pipeline.Request`.
Such request is will be used to call user handler with all the rules
applied. An axion`s `pipeline.Response` will later on be used to
create, using respone processer, a framework's response.

Partially: #190